+++ extensions/ql-vscode/src/view/common/TextButton.tsx
@@ -3,12 +3,12 @@ import { styled } from "styled-components";
 
 type Size = "x-small" | "small" | "medium" | "large" | "x-large";
 
-const StyledButton = styled.button<{ size: Size }>`
+const StyledButton = styled.button<{ $size: Size }>`
   background: none;
   color: var(--vscode-textLink-foreground);
   border: none;
   cursor: pointer;
-  font-size: ${(props) => props.size ?? "1em"};
+  font-size: ${(props) => props.$size ?? "1em"};
   padding: 0;
 `;
 
@@ -26,7 +26,7 @@ const TextButton = ({
   children: React.ReactNode;
 }) => (
   <StyledButton
-    size={size}
+    $size={size}
     onClick={onClick}
     className={className}
     title={title}