+++ extensions/ql-vscode/CHANGELOG.md
@@ -1,5 +1,7 @@
 # CodeQL for Visual Studio Code: Changelog
 
+## [UNRELEASED]
+
 ## 1.9.4 - 6 November 2023
 
 No user facing changes.