+++ extensions/ql-vscode/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [UNRELEASED]
 
+- The "Sort by Language" action in the databases view now sorts by name within each language. [#3055](https://github.com/github/vscode-codeql/pull/3055)
+
+## 1.9.4 - 6 November 2023
+
+No user facing changes.
+
 ## 1.9.3 - 26 October 2023
 
 - Sorted result set filenames now include a hash of the result set name instead of the full name. [#2955](https://github.com/github/vscode-codeql/pull/2955)